+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 #### What's Improved
 - Playback bar now has a marquee effect
+- Play/pause button now changes from square to circle depending on the state
 - Added a way to access the system equalizer from the playback menu.
 
 #### What's Fixed